Flatten down the db/models;
we do not need them separate this is
a little simpler and sufficient

remove unique constraints on states lol
