+++ backend/ballotnav/api/dropoffs.py
@@ -4,28 +4,22 @@ import structlog
 from webargs import fields
 from webargs.flaskparser import use_args
 
-from ballotnav.models import DropoffModel, dropoff_schema, dropoffs_schema
+from ballotnav.models import Dropoff, dropoff_schema
 
 dropoffs = Blueprint("dropoffs", __name__)
 logger = structlog.get_logger()
 
 
 @dropoffs.route("/dropoffs", methods=["GET"])
-@use_args(
-    {
-        "state_name": fields.Str(),
-        "county": fields.Str(),
-    }, location="query"
-)
+@use_args({"state_name": fields.Str(), "county": fields.Str()}, location="querystring")
 def get_location(args):
     """ get location """
-    logger.msg(f"Get location for")
-    res = DropoffModel(app=current_app).get(args)
+    res = Dropoff.get(dict(county=args.get("county"), state_short_code=args.get("state_name")))
 
     if not res:
         logger.msg("No locations found", args=args)
-        return res, 404
-    return {"dropoffs": dropoffs_schema.dump(res)}, 200
+        return {"dropoffs": [], "status": "error"}, 404
+    return {"dropoffs": dropoff_schema.dump(res, many=True)}, 200
 
 
 @dropoffs.route("/dropoff", methods=["POST"])
@@ -49,8 +43,8 @@ def create_location(args):
     """ create a dropoff location entry """
     logger.info(f"CREATE a location call")
 
-    res = DropoffModel(current_app).create(args)
+    res = Dropoff.create(args)
 
     if not res:
         return {"status": "error"}, 400
-    return dropoff_schema.dump(res), 201
+    return {"status": "ok", "dropoffs": {**dropoff_schema.dump(res)}}, 201
