--- backend/ballotnav/db/__init__.py
@@ -1,2 +1,0 @@
-""" db """
-from .dropoff import Dropoff
